Moving a few internal classes to android.internal package.
