+++ formula-android/src/main/java/com/instacart/formula/android/internal/ActivityManager.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.Lifecycle
 import com.instacart.formula.activity.ActivityResult
 import com.instacart.formula.fragment.FragmentEnvironment
 import com.instacart.formula.android.ActivityStore
-import com.instacart.formula.integration.FragmentFlowRenderView
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.disposables.Disposable
