+++ formula-android/src/main/java/com/instacart/formula/android/internal/FormulaFragmentViewFactory.kt
@@ -1,9 +1,11 @@
-package com.instacart.formula.integration
+package com.instacart.formula.android.internal
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.instacart.formula.android.FeatureView
 import com.instacart.formula.android.ViewFactory
+import com.instacart.formula.integration.FeatureEvent
+import com.instacart.formula.integration.FragmentId
 import java.lang.IllegalStateException
 
 internal class FormulaFragmentViewFactory(
